+++ tests/test_visitors/test_ast/test_complexity/test_overuses/test_overused_string.py
@@ -1,6 +1,6 @@
 import pytest
 
-from wemake_python_styleguide.compat.constants import PY314
+from wemake_python_styleguide.compat.constants import PY312, PY314
 from wemake_python_styleguide.violations.complexity import (
     OverusedStringViolation,
 )
@@ -168,6 +168,100 @@ def fourth():
     """Docs."""
     {0}
 '''
+
+# See:
+# https://github.com/wemake-services/wemake-python-styleguide/issues/3805
+module_attribute_docstrings = """
+first = 1
+{0}
+
+second: int = 2
+{0}
+
+third: int
+{0}
+"""
+
+class_attribute_docstrings = """
+class Some:
+    first = 1
+    {0}
+
+    second: int = 2
+    {0}
+
+    def __init__(self):
+        self.third = 3
+        {0}
+"""
+
+not_an_attribute_docstring = """
+def first():
+    x = y = 1
+    {0}
+
+def second():
+    x, y = call()
+    {0}
+
+def third():
+    y: int = 0
+    {0}
+
+def fourth():
+    x = 1
+    {0}
+"""
+
+# `x.some = 1` defines an attribute of `x`, not of the module, the class,
+# or the instance we are in. So, there's nothing here to document.
+foreign_attribute_docstrings = """
+x.some = 1
+{0}
+
+def first():
+    x.some = 1
+    {0}
+
+class Some:
+    def second(self, x):
+        x.some = 1
+        {0}
+
+    def third(self):
+        self.some.other = 1
+        {0}
+"""
+
+# Type aliases are documented the very same way, see:
+# https://discuss.python.org/t/docstrings-for-type-aliases/108901
+explicit_type_alias_docstrings = """
+Timeout: TypeAlias = float
+{0}
+
+Retries: TypeAlias = int
+{0}
+
+Backoff: TypeAlias = float
+{0}
+"""
+
+type_alias_docstrings = pytest.param(
+    """
+    type Timeout = float | None
+    {0}
+
+    type Retries = int
+    {0}
+
+    type Backoff = float
+    {0}
+    """,
+    marks=pytest.mark.skipif(
+        not PY312,
+        reason='`type` aliases are only in Python 3.12+',
+    ),
+)
 
 EXPECTED_LOCATION = (2, 8)
 
@@ -373,6 +467,10 @@ def test_common_strings_allowed(
     [
         module_docstring,
         function_docstrings,
+        module_attribute_docstrings,
+        class_attribute_docstrings,
+        explicit_type_alias_docstrings,
+        type_alias_docstrings,
     ],
 )
 @pytest.mark.parametrize(
@@ -400,20 +498,29 @@ def test_docstrings_not_counted(
 
 
 @pytest.mark.parametrize(
+    'strings',
+    [
+        not_a_docstring,
+        not_an_attribute_docstring,
+        foreign_attribute_docstrings,
+    ],
+)
+@pytest.mark.parametrize(
     'string_value',
     [
         '"""Not a docstring."""',
         '"Not a docstring."',
     ],
 )
-def test_strings_after_docstrings_counted(
+def test_strings_in_other_places_counted(
     assert_errors,
     parse_ast_tree,
     options,
+    strings,
     string_value,
 ):
-    """Ensures that only the first string in a body is a docstring."""
-    tree = parse_ast_tree(not_a_docstring.format(string_value))
+    """Ensures that strings documenting nothing are still counted."""
+    tree = parse_ast_tree(strings.format(string_value))
 
     option_values = options(max_string_usages=0)
     visitor = StringOveruseVisitor(option_values, tree=tree)